refactor network configs into core to prevent circular deps
